[efi] Add TPM measurement API via TCG v1 and TCG v2 EFI protocols
This follows the TCG EFI firmware profile spec and provides an api for the rest of ipxe to measure a given memory segment into a specified tpm pcr.
Also allows to log a separator event at the end of ipxe measurements.
Absence of a TPM or a disabled TPM is not treated as an error.
Measure the kernel and kernel commandline.

Signed-off-by: Vishnu Rangayyan <vrangayyan@nvidia.com>